Added a refresh button
Added a refresh button to manually refresh the package list.  The list
it normally updated whenever the window is opened, but some users may
keep the window open for a long period of time and in the meantime some
new packages may have been released.  It is easier to click the Refresh
button than closing and reopening the window.